backend/ai-agent/requirements.txt
annotated-types
attrs
blinker
cachetools
certifi
charset-normalizer
click
flasgger
Flask
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-generativeai
googleapis-common-protos
grpcio
grpcio-status
httplib2
idna
itsdangerous
Jinja2
jsonschema
jsonschema-specifications
MarkupSafe
mistune
packaging
proto-plus
protobuf
pyasn1
pyasn1_modules
pydantic
pydantic_core
pyparsing
python-dotenv
PyYAML
referencing
requests
rpds-py
rsa
six
tqdm
typing-inspection
typing_extensions
uritemplate
urllib3
Werkzeug
